Improved makefile to hide errors and cleanup version string

+++ Makefile
@@ -110,14 +110,14 @@ res/$(1): themes/$(1)/* #themes/$(1)/font/* themes/$(1)/font7x6/* themes/$(1)/ti
 	@$(ECHO) "Building theme $(1)..."
 	@mkdir -p res
 	@mkdir -p themetemp
-	@-cp -f themes/$(1)/colors.txt themetemp
-	@-cp -f themes/$(1)/bevel.* themetemp
-	@-cp -f themes/$(1)/vu.* themetemp
-	@-cp -f themes/$(1)/analyzor.* themetemp
-	@-cp -f themes/$(1)/logo.* themetemp
-	@-cp -f themes/$(1)/catometer.* themetemp
-	@-cp -f themes/$(1)/cursor.* themetemp
-	@-cp -f themes/$(1)/icon.* themetemp
+	-@if test -e themes/$(1)/colors.txt; then cp -f themes/$(1)/colors.txt themetemp ; fi
+	-@if test -e themes/$(1)/bevel.*; then cp -f themes/$(1)/bevel.* themetemp ; fi
+	-@if test -e themes/$(1)/vu.*; then cp -f themes/$(1)/vu.* themetemp ; fi
+	-@if test -e themes/$(1)/analyzor.*; then cp -f themes/$(1)/analyzor.* themetemp ; fi
+	-@if test -e themes/$(1)/logo.*; then cp -f themes/$(1)/logo.* themetemp ; fi
+	-@if test -e themes/$(1)/catometer.*; then cp -f themes/$(1)/catometer.* themetemp ; fi
+	-@if test -e themes/$(1)/cursor.*; then cp -f themes/$(1)/cursor.* themetemp ; fi
+	-@if test -e themes/$(1)/icon.*; then cp -f themes/$(1)/icon.* themetemp ; fi
 	-@if test -d themes/$(1)/font; then $(MAKEBUNDLE) themetemp/8x8.fnt themes/$(1)/font ; fi
 	-@if test -d themes/$(1)/font7x6; then $(MAKEBUNDLE) themetemp/7x6.fnt themes/$(1)/font7x6 ; fi
 	-@if test -d themes/$(1)/tiny; then $(MAKEBUNDLE) themetemp/4x6.fnt themes/$(1)/tiny ; fi
@@ -145,7 +145,7 @@ src/version_number.h: src/version
 	@echo '#ifndef VERSION_NUMBER' > src/version_number.h
 	@echo '#define VERSION_NUMBER' >> src/version_number.h
 	@echo -n '#define VERSION "' >> src/version_number.h
-	@echo -n `cat src/version` >> src/version_number.h
+	@echo -n `cat src/version | tr -d '\r\n'` >> src/version_number.h
 	@echo '"' >> src/version_number.h
 	@echo '#endif' >> src/version_number.h
